requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.0.7
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-filter==2.2.0
django-formtools==2.2
django-import-export==2.0.2
django-js-asset==1.2.2
django-mdeditor==0.1.17
django-mptt==0.11.0
django-redis==4.11.0
django-reversion==3.0.7
django-simpleui==3.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.25.1
future==0.18.2
greenlet==0.4.15
httplib2==0.9.2
importlib-metadata==1.5.0
jdcal==1.4.1
kombu==4.6.8
Markdown==3.1.1
MarkupPy==1.14
monotonic==1.5
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
Pygments==2.5.2
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.3
qcloudsms-py==0.1.4
redis==3.4.1
six==1.14.0
tablib==1.1.0
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
